+++ src/gpu/GrIndexBuffer.h
@@ -21,7 +21,7 @@ public:
      * @return the maximum number of quads using full size of index buffer.
      */
     int maxQuads() const {
-        return this->sizeInBytes() / (sizeof(uint16_t) * 6);
+        return static_cast<int>(this->sizeInBytes() / (sizeof(uint16_t) * 6));
     }
 protected:
     GrIndexBuffer(GrGpu* gpu, bool isWrapped, size_t sizeInBytes, bool dynamic, bool cpuBacked)
