+++ src/gpu/GrRenderTarget.cpp
@@ -57,7 +57,7 @@ void GrRenderTarget::resolve() {
 }
 
 size_t GrRenderTarget::sizeInBytes() const {
-    int colorBits;
+    size_t colorBits;
     if (kUnknown_GrPixelConfig == fDesc.fConfig) {
         colorBits = 32; // don't know, make a guess
     } else {
